+++ packages/expo-ui/ios/Modifiers/ViewModifierRegistry.swift
@@ -1118,23 +1118,6 @@ internal struct ListSectionMargins: ViewModifier, Record {
   }
 }
 
-internal enum AxisOptions: String, Enumerable {
-  case horizontal
-  case vertical
-  case both
-
-  func toAxis() -> Axis.Set {
-    switch self {
-    case .vertical:
-      return .vertical
-    case .horizontal:
-      return .horizontal
-    case .both:
-      return [.vertical, .horizontal]
-    }
-  }
-}
-
 internal struct GridCellUnsizedAxes: ViewModifier, Record {
   @Field var axes: AxisOptions?